requirements.txt
discord.py
requests
beautifulsoup4
fake_headers
prettytable
python-decouple